requirements.txt
numpy 
pandas 
scikit-learn 
streamlit 
matplotlib 
seaborn 
google-cloud-storage
pyyaml
